move composable to its folder
