+++ app/src/main/java/com/sanaa/tudee_assistant/presentation/screen/taskScreen/TasksScreen.kt
@@ -40,6 +40,7 @@ import androidx.compose.ui.unit.dp
 import com.sanaa.tudee_assistant.R
 import com.sanaa.tudee_assistant.presentation.composable.CustomDatePickerDialog
 import com.sanaa.tudee_assistant.presentation.composable.DeleteComponent
+import com.sanaa.tudee_assistant.presentation.composable.TaskStatusTabs
 import com.sanaa.tudee_assistant.presentation.design_system.component.DayItem
 import com.sanaa.tudee_assistant.presentation.design_system.component.TudeeSnackBar
 import com.sanaa.tudee_assistant.presentation.design_system.component.button.FloatingActionButton
